+++ include/workarounds/simplehttp.hpp
@@ -1,4 +1,0 @@
-#pragma once
-#include "mruby.h"
-
-void workarounds_simplehttp(mrb_state*);