+++ scripts/mruby/darwin.rb
@@ -30,5 +30,6 @@ MRuby::CrossBuild.new('x86_64-apple-darwin19') do |conf|
   conf.gem github: 'Asmod4n/mruby-uri-parser'
   conf.gem github: 'matsumotory/mruby-simplehttp'
   conf.gem github: 'matsumotory/mruby-simplehttpserver'
+  conf.gem github: 'mattn/mruby-base64'
   conf.gem github: 'mattn/mruby-require'
 end