+++ scripts/mruby/emscripten.rb
@@ -26,5 +26,6 @@ MRuby::CrossBuild.new('emscripten') do |conf|
   conf.gem github: 'matsumotory/mruby-simplehttp'
   # I dunno if it ever makes sense to enable this for web?
   #conf.gem github: 'matsumotory/mruby-simplehttpserver'
+  conf.gem github: 'mattn/mruby-base64'
   conf.gem github: 'mattn/mruby-require'
 end